refactor: enumerate -> enums
