+++ src/main/kotlin/org/team14/webty/review/dto/ReviewRequest.kt
@@ -2,7 +2,7 @@ package org.team14.webty.review.dto
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import org.springframework.web.multipart.MultipartFile
-import org.team14.webty.review.enumrate.SpoilerStatus
+import org.team14.webty.review.enums.SpoilerStatus
 
 data class ReviewRequest(
     val webtoonId: Long,
@@ -10,5 +10,5 @@ data class ReviewRequest(
     val title: String,
     val spoilerStatus: SpoilerStatus,
     @JsonIgnore
-	var images: List<MultipartFile>? = ArrayList() // 이미지 파일 리스트 추가
+    var images: List<MultipartFile>? = ArrayList() // 이미지 파일 리스트 추가
 )
