+++ src/main/kotlin/org/team14/webty/webtoon/repository/WebtoonRepository.kt
@@ -7,19 +7,21 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
 import org.team14.webty.webtoon.entity.Webtoon
-import org.team14.webty.webtoon.enumerate.Platform
-import java.util.Optional
+import org.team14.webty.webtoon.enums.Platform
+import java.util.*
 
 @Repository
 interface WebtoonRepository : JpaRepository<Webtoon, Long> {
 
-    @Query("""
+    @Query(
+        """
         SELECT w FROM Webtoon w 
         WHERE (:webtoonName IS NULL OR w.webtoonName LIKE %:webtoonName%) 
         AND (:platform IS NULL OR w.platform = :platform) 
         AND (:authors IS NULL OR w.authors LIKE %:authors%) 
         AND (:finished IS NULL OR w.finished = :finished)
-    """)
+    """
+    )
     fun searchWebtoons(
         @Param("webtoonName") webtoonName: String?,
         @Param("platform") platform: Platform?,
